Added some CSS to the templates and modified census request api
Added CSS to both state.html and county.html, added bootstrap to them. Allowed census to take in both a state and county fip code. TODO: add error checking to the state county input to ensure its a fips code and make sure that the fips code is a string.